Add n_rows to VRUDecomposition ctor, still not working
For kernel persistence, the matrix is no longer square,
the number of rows must be given explicitly, so that
the pivots vector is of the right size.

Also changed all cout to cerr, to guarantee that we see
everything printed out before the program crashes.

After fixing this, there is still some malloc corrupted
top error.